+++ docs/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Releases
 
-## v1.45.0
+## v1.46.0
 
 Released: ???
+
+Features:
+- Added more Galatea fishing drops for potential future update to CT 3.0.
+- Renamed The Sea Emperor to The Loch Emperor everywhere.
+- Deleted outdated feature "Highlight matching items in Attribute Fusion" because there is no old attribute fusion anymore.
+- Deleted outdated feature "Price per T1 attribute shard" because there is no high level Attribute Shards on AH.
+
+Bugfixes:
+- Fixed Flake the Fish rarity (displayed as Special now).
+
+Other:
+-
+
+## v1.45.0
+
+Released: 2025-06-23
 
 Features:
 - Worm Profit tracker - in Gemstone chambers mode, the price of a Gemstone Mixture is subtracted from the price of a Gemstone Chamber, for more accurate profits. You can configure buy method for mixtures in the settings.